+++ examples/complete/data.tf
@@ -55,33 +55,6 @@ data "aws_iam_policy_document" "access_logs_bucket" {
 
 data "aws_elb_service_account" "main" {}
 
-data "aws_iam_policy_document" "task_execution_role_policy_doc" {
-  #checkov:skip=CKV_AWS_111:Ensure IAM policies does not allow write access without constraints"
-  statement {
-    effect = "Allow"
-    actions = [
-      "logs:CreateLogStream",
-      "logs:PutLogEvents",
-    ]
-
-    resources = ["arn:${local.partition}:logs:::log-group:${local.name}"]
-  }
-
-  statement {
-    effect = "Allow"
-    actions = [
-      "ecr:GetAuthorizationToken",
-      "ecr:BatchCheckLayerAvailability",
-      "ecr:GetDownloadUrlForLayer",
-      "ecr:BatchGetImage",
-      "logs:CreateLogStream",
-      "logs:PutLogEvents"
-    ]
-
-    resources = ["*"]
-  }
-}
-
 data "aws_vpc" "supporting" {
   filter {
     name   = "tag:Name"
@@ -115,4 +88,8 @@ data "aws_subnet" "private" {
 
 data "aws_ecs_cluster" "ecs" {
   cluster_name = local.supporting_resources_name
+}
+
+data "aws_kms_alias" "supporting_kms" {
+  name = "alias/${local.supporting_resources_name}"
 }