+++ tests/supportingResources/main.tf
@@ -17,24 +17,28 @@ module "ecs_vpc" {
 
 module "kms_key" {
   source                  = "boldlink/kms/aws"
+  version                 = "1.1.0"
   description             = "A test kms key for ecs cluster"
   create_kms_alias        = true
   alias_name              = "alias/${local.name}"
   enable_key_rotation     = true
+  kms_policy              = local.kms_policy
   deletion_window_in_days = 7
   tags                    = local.tags
 }
 
 resource "aws_cloudwatch_log_group" "cluster" {
-  #checkov:skip=CKV_AWS_158:Ensure that CloudWatch Log Group is encrypted by KMS"
-  name              = "${local.name}-log-group"
-  retention_in_days = 0
+  name              = "/aws/ecs-cluster/${local.name}-log-group"
+  retention_in_days = 1
+  kms_key_id        = module.kms_key.arn
   tags              = local.tags
 }
 
+
 module "cluster" {
-  source = "git::https://github.com/boldlink/terraform-aws-ecs-cluster.git?ref=1.0.1"
-  name   = local.name
+  source  = "boldlink/ecs-cluster/aws"
+  version = "1.0.1"
+  name    = local.name
   configuration = {
     execute_command_configuration = {
       kms_key_id = module.kms_key.key_id